Final.h addition
Moved definitions to a new class. Added comments. Changed how PID
reacted.
